Optimize file handling and update version
- Enhanced DDS file handling in MenuArtConverter.cs by adding a check to bypass conversion for files already in DDS format.
- Improved file conversion logic in PictoConverter.cs to include a DDS check, mirroring the optimization in MenuArtConverter.cs, and adjusted control flow for .xtx file deletion.
- Adjusted the data type of `TrackId` from `int` to `long` in KaraokeTape.cs because some jsons add .0 to numbers.
-  Version bump

+++ JustDanceEditor.Converter/Converters/Images/MenuArtConverter.cs
@@ -53,6 +53,13 @@ public static class MenuArtConverter
     {
         string outputPath = Path.Combine(tempFolder, fileName + ".dds");
 
+        // If the file starts with "DDS ", it's already a DDS file and we simply rename it
+        if (File.ReadAllBytes(inputPath).Take(4).SequenceEqual("DDS "u8.ToArray()))
+        {
+            File.Move(inputPath, outputPath);
+            return outputPath;
+        }
+
         // Convert the XTX file to DDS
         XTXExtractAdapter.ConvertToDDS(inputPath, outputPath);
         File.Delete(inputPath);
