+++ JustDanceEditor.Converter/Converters/Images/PictoConverter.cs
@@ -44,12 +44,20 @@ public static class PictoConverter
             stream.Close();
             newStream.Close();
 
-            // Run xtx_extract on the new file, parameters: -o {filename}.dds {filename}.xtx
-            // Print the output to the console
-            XTXExtractAdapter.ConvertToDDS(Path.Combine(convert.TempPictoFolder, fileName + ".xtx"), Path.Combine(convert.TempPictoFolder, fileName + ".dds"));
+            // If the file starts with "DDS ", it's already a DDS file and we simply rename it
+            if (File.ReadAllBytes(Path.Combine(convert.TempPictoFolder, fileName + ".xtx")).Take(4).SequenceEqual("DDS "u8.ToArray()))
+            {
+                File.Move(Path.Combine(convert.TempPictoFolder, fileName + ".xtx"), Path.Combine(convert.TempPictoFolder, fileName + ".dds"));
+            }
+            else
+            {
+                // Run xtx_extract on the new file, parameters: -o {filename}.dds {filename}.xtx
+                // Print the output to the console
+                XTXExtractAdapter.ConvertToDDS(Path.Combine(convert.TempPictoFolder, fileName + ".xtx"), Path.Combine(convert.TempPictoFolder, fileName + ".dds"));
 
-            // Delete the .xtx file
-            File.Delete(Path.Combine(convert.TempPictoFolder, fileName + ".xtx"));
+                // Delete the .xtx file
+                File.Delete(Path.Combine(convert.TempPictoFolder, fileName + ".xtx"));
+            }
 
             // Convert the .dds file to .png
             Image<Bgra32> newImage;
